+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### 🚀 Features
 
 - *(cli)* Handle multiple instances of filecrabs
+- *(cli)* Add a command to `add` instances to a filecrab config
 
 ## [0.1.1] - 2024-04-22